Track early return in wasm_generator
